Updated hitlag extension to ceil on ApplyHitlag

+++ pkg/enemy/enemy.go
@@ -48,8 +48,8 @@ type Enemy struct {
 	mods []modifier.Mod
 
 	//hitlag stuff
-	timePassed   float64
-	frozenFrames float64
+	timePassed   int
+	frozenFrames int
 	queue        []queue.Task
 }
